Many minor edits to comments in libglimmer; no answer changes.
This is the first of several cosmetic commits this morning. I'm going one directory at a time.

(1) I changed 'Glimmer-CISM' to 'CISM' in the headers.
(2) I changed '2005-2013' to '2005-2014' in the headers.
(3) I removed the string '*FD' wherever it appeared in comments.
    (Here, the * symbol is taken literally, not as a wildcard.)
(4) I cleaned up some 'TODO' comments.  However, most of the TODOs aren't done.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1821 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: a54645cba0f511987b736265d42155c783949724