Commenting out i/o variables lat and lon from glide_vars.def
These variables are not being used anywhere in the code and are confusing to have as i/o variables.  It appears the functionality exists in CISM to calculate lat and lon fields from user-supplied projection information to populate these as output fields (which could be useful), but additional work would be needed to determine exactly how to do that and where.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1870 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: dca72225adc61eb6a487151c6e50cc4c776850d1