I added a new nonlinear convergence option for the Glissade solver, which_ho_resid = HO_RESID_L2NORM_RELATIVE = 4.
If this option is set in the config file, Glissade will	determine that
the nonlinear velocity solve has converged when	|Ax - b| / |b| is below
a user-defined threshold.

The default remains HO_RESID_L2NORM = 3, in which case the solver converges
when |Ax - b| is below a user-defined threshold	(10^-4 by default).

I set the relative threshold to 10^(-7) so that standard test cases (dome, shelf)
converge in about the same number of iterations	as with the default option.

The relative option may	be more appropriate for large problems (e.g., whole ice	sheets),
where the number of DOFs is much larger than for test cases, and it may be unnecessarily
strict to require convergence of |Ax - b| to a factor of 10^-4.

Since the default has not changed, there are no answer changes (confirmed by Mac LIVV).

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1884 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 114536c1924348344e74081f0f1558cd84de6288