Modifications to subroutine glissade_calcbmlt in glissade_temp.F90.
There were a couple of issues here.  First, there was a	missing	time term in the basal
melt calculation.  I added this	term, with the result that basal melt (or freeze-on)
is greater in magnitude than before.

Second,	for the case that T_b = T_pmp and bmlt < 0, the	code was computing freeze-on
even when no basal water was present.  To prevent this from becoming a steady but
erroneous source of basal growth, I added some code to set T_b to a value slightly
below T_pmp when this occurs.  Then on the next timestep, the code will solve for
T_bed in such a	way that fluxes balance	and no further freeze-on occurs.

In our standard test cases, these changes make no difference to the dynamics,
and the results remain BFB for Mac LIVV.

This subroutine does not exactly conserve energy.  We should revisit this at some point.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1890 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: ef0f5a7e6c8ab4caa98ee78341796073c18f543a