I modified the basal sliding friction term, bfricflx, that contributes to the energy balance at the bed.  The Glissade velocity solver has been computing this value in a physically consistent way since revision 1807, but previously this value was not being applied in glissade_temp.
This commit changes answers for time-dependent problems with Glissade!  (E.g., the evolving dome)
This is the first of several answer-changing commits.

The changes are small but nonzero for the evolving dome (around the 8th decimal place for velocity, not visible to the eye),
because the basal sliding is very slow by default (which_ho_babc = 4).  Changes will be larger
for problems with faster sliding.

Answers are BFB for diagnostic tests (confirmed by Mac LIVV).

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1892 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 550e837456b849c89183c9dd3c13d4d0ea956620