Changes in internal heat dissipation: I changed the default value of which_ho_disp from HO_DISP_SIA = 0 to HO_DISP_FIRSTORDER = 1.
The new default is more physically consistent.  Internal dissipation is now proportional
to the tensor product of the stress and strain rate, consistent with the approximation
(BP, L1L2, SSA or SIA).

This commit changes answers for time-dependent problems with Glissade.
For the evolving dome with BP dynamics, the ice is slightly colder after 10 years,
and speeds are ~1% slower.

Answers are unchanged for diagnostic problems (confirmed by Mac LIVV).

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1893 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 7487ac08b1797c3d4889726f9412bc841d069974