In the incremental remapping transport scheme, I changed the default value of dp_midpt from .false. to .true.
This change slightly improves accuracy in computing departure regions.
The default was set to .false. a while back for comparisons to an older version of IR,
and until now was not updated.

Answers change for time-dependent problems using Glissade, in particular the evolving dome.

Answers for diagnostic problems remain unchanged (confirmed by Mac LIVV).

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1894 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 285f0d769c47ca0dcfeac90c6e42045d011d4d73