I changed the default value of which_ho_gradient_margin from HO_GRADIENT_MARGIN_ICE_ONLY = 2 to HO_GRADIENT_MARGIN_ICE_LAND = 1.
With the old default, only ice-covered cells were included when computing the surface elevation gradient.
As a result, gradients tend to be very small near land-terminating margins.

With the new default, both ice-covered cells and ice-free land cells are included in the gradient.
This change improves accuracy for problems with a land-terminating margin (e.g., dome and Halfar).
The initial velocities are faster near the margin, and the ice spreads out more quickly.
In the LIVV diagnostics, the sharp edges in the velocity pattern near the margin have disappeared;
the velocity is now much smoother.

Answers are unchanged for shelf problems (circular, confined, Ross) and problems with no ice margin (ISMIP-HOM).

At this point the diagnostic dome and evolving dome tests are no longer BFB with Mac LIVV.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1895 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: a01360e61be1f7dc8a335bde0fbd3c9a93ec9411