I consolidated two equations (of the form y = x^p) into one in the computation of effective viscosity.
This results in changes at machine roundoff level.

The reason for the change is that this calculation is done many times during matrix assembly,
so assembly is measurably faster with the change.

I expected all the test results to change at machine roundoff level.
However, the circular shelf and ISHOM A and C cases are still BFB according to Mac LIVV.
The confined shelf is no longer BFB, but changes are very small.

The diagnostic and evolving dome were already no longer BFB from previous commits.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1896 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 4a70514455b977b1f8205081a9173ca4e9b22435