More minor edits for Glissade modules.
The biggest change is that I added a subroutine to initialize the Trilinos solve
with the solution from the previous timestep.  Previously, Trilinos was starting
each timestep with an assumption of zero velocity.  This changes Trilinos
answers at the solver roundoff level.

I verified that Trilinos still converges to the same solution (within solver roundoff)
as the native PCG solver for the evolving dome.

Answers are unchanged for the standard tests (confirmed by Mac LIVV).

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1906 42a2fd4b-9c52-44de-a8a9-1772c67210cd