Moved heat dissipation and basal friction calculations; replaced pointers with subroutine arguments
Answers for standard problems are unchanged (verified by Mac LIVV for evolving dome).

Previously, the interior heat dissipation was computed in glissade_therm for both Glam and Glissade.
Basal friction was computed in glissade_therm for Glam (but was already part of the velocity solution
for Glissade).

With this commit, the call to the heat dissipation subroutine moves from glissade_therm.F90 to glissade.F90,
just after the velocity solve.  This means that for exact restart, the 3D dissip field is now required,
but the 3D fields of tau_eff and efvs are not.  I changed the restart variable list accordingly
(in glide_setup.F90) and verified that the model restarts exactly for dome and EISMINT-2 cases.
In glide_vars.def, dissip and dissipstag are now loadable; tau_eff and efvs are not.
The subroutines that compute heat dissipation remain in glissade_therm.

The basal friction calculation remains in the velocity solver for Glissade.
For Glam, it has moved to glissade.F90, just after the velocity solve.
The relevant Glam subroutine, glam_basal_friction, is now in glam_velo.F90.

The units of dissip are now deg/s, consistently throughout the code.  Previously the units
switched between deg and deg/s.

In glam_velo.F90, I noticed that the call to glide_calcstrsstr (required for computing tau_eff)
took place after the Picard nonlinear solve but not the JFNK solve.  I moved the call so that
it takes place for either nonlinear solver.  This changes the answers for Glam evolving problems
using JFNK.

I decided to pass a number of arguments in and out of glissade_therm_driver rather than use pointers
to the model derived types.  This means	that glissade_therm_driver can use SI units without
having to rescale fields within	  the subroutine. (Input arguments can be passed with SI units.)
For now, glissade_therm_driver still uses model	units for time and thickness, but I will
switch entirely	to SI with a future commit (which will change answers at roundoff level).