Removed some scaling constants from glissade_therm.F90.
I converted three more subroutines (glissade_flow_factor and two heat
dissipation subroutines) to SI units.  This means that when these subroutines
are called from glissade_tstep, input arguments must be passed with SI units,
and in one case, output must be scaled back to model units.  The net effect
is to move scaling constants up a level, making glissade_tstep a bit messier
but glissade_therm cleaner.  Since glissade_therm is the module to be translated
to MPAS, I'm trying to make it as clean as possible.

With this commit, answers change at the machine roundoff level (around the 13th
decimal place) for the evolving dome and EISMINT-2 problems.