Implemented a small correction to basal melting for the enthalpy scheme.
Previously, the basal melt rate (bmlt) was always computed as

     bmlt = bflx / (rhoi*Li)

where bflx is the net heat flux available for melting, rhoi is ice density,
and Li is the latent heat of melting.

For the enthalpy scheme, I changed the denominator to (rhoi*Li - enth(upn)),
where enth(upn) is the specific enthalpy of the lowest ice layer.
This results in a slightly faster melt rate for temperate ice
(because of internal melting that has already occurred) and a
slightly slower rate for cold ice (because of the extra energy needed
to raise the temperature to T_pmp).  Note that enth = 0 for T = Tpmp
and waterfrac = 0, in which case the denominator reduces to rhoi*Li.

This change will make it easier to conserve energy exactly.
Some mods would be needed, however, to conserve energy in the rare case that
the basal melting extends beyond the lowest ice layer.

Basal melt rates for EISMINT-2 (with the enthalpy scheme) change by a fraction
of a percent. Since basal melt rates are ignored in the continuity equation by default,
ice state variables are unchanged.