Minor cleanup after bug fixes
In glissade_temp.F90, I removed the logical parameter that could be toggled
to run either with or without the bug fixes. Now, bug fixes are always applied.

Also, I verified that Glissade aborts if the user supplies a restart file
with 'temp' instead of 'tempstag' as an input field.  This didn't require
any code changes.

Answers are unchanged compared to the previous commit.