Changed some variable and subroutine names
In several files I changed "flow_fudge_factor" to "flow_enhancement_factor", given that
there are legitimate physical reasons for values greater than 1, so it is not simply a fudge.

In glissade_therm.F90, I changed the names of the two subroutines that
compute matrix elements for the vertical solve.

I removed a couple of debug print statements in glide_types.F90.