Add tempstag_pmp as an output variable
This creates a new internal model variable called temp_pmp that is the
pressure melting point temperature.  Right now it is calculated in
glissade_diagnostic_variable_solve.  It is *not* calculated in Glide.

There is an output variable called tempstag_pmp that can be used to
output this field.  There is not a temp_pmp output variable added since
it is not being calculated in Glide.