Update ISMIP-HOM plot script.
This changes the ISMIP-HOM plot script to plot the newest *set* of CISM
output files found in the output files directory.

`runISMIP_HOM.py` produces filenames following this pattern:

```
ismip-hom-?[-MOD].RESO.[pPROC].EXT
```

where `?` is the experiment (e.g. `a`), `[-MOD]` is a optionally user
specified filename modifier, `RESO` is the size of the experiment,
`[pPROC.]` is the number of processors the tests was run with
(optional), and `EXT` is the file extension.

A set of files is all the `?` and `RESO` files grouped by the [-MOD] and
[pPROC.] identifies (of lack thereof).

If more than one set is detected, only the newest set is plotted. Using
the `-f` option, a specific set can be selected by passing
`plotISMIP_HOM.py` any output file that belongs to the set.

A sizes option can be passed which may *limit* the set of sizes plotted
within a set.