Added a depth-dependent bmlt_float option
After some testing, I replaced the recently added (but little tested) BMLT_FLOAT_CAVITY_THCK option
with a new BMLT_FLOAT_DEPTH option, whichbmlt_float = 3. The new option allows the user to apply
a bmlt_float field that is a piecewise linear function of depth, with the following form:

- Above zfrzmax, bmlt_float = -frzmax
- Between zfrzmax and zmelt0, the freezing rate ramps down linearly to 0.
- Between zmelt0 and zmeltmax, the melt rate ramps up linearly to meltmax.
- Below zmeltmax, bmlt_float = meltmax

Here, z < 0 by definition, with zfrzmax > zmelt0 > zmeltmax.

Default values are zfrzmax = -100 m, zmelt0 = -200 m, zmeltmax = -500 m,
frzmax = 0, meltmax = 10 m/yr.

I tested this option with various parameter settings for Greenland simulations.

- I found that the Petermann and 79 N ice shelves require frzmax of ~2 to 3 m/yr
to avoid retreat of the calving front due to negative SMB. (Possibly, this freezing rate
is not realistic but is compensating for a negative SMB bias.)

- I found that a large value of meltmax (~100 m/yr) is required to prevent
the Petermann GL from advancing as the upstream glacier velocity spins up
in the first few thousand years of the simulation.

Also, I switched the sign of the local BMB diagnostic in the log file, and modified
the local acab and bmlt diagnostics to be based on acab_applied and bmlt_applied
instead of acab and bmlt.

I fixed a comment in glissade_calving.

Answers are BFB except for simulations with whichbmlt_float = 3.